Added files via upload
Added a prototype automatic email setup to the application submission page. Also added input fields for comments and score changing on the score table page, and added a link to the incomplete nominees page as per the project description. Full list of changes is in the changelog included in the zipped folder.